refactor cli run into AppLauncher class
